Refactor action package to control
