+++ control/controller.go
@@ -1,4 +1,4 @@
-package action
+package control
 
 import (
 	"encoding/json"
