+++ control/doc.go
@@ -1,0 +1,6 @@
+/*
+Package control handles incoming requests and actions.
+
+The Controller type wraps a LaunchManager and implements HTTP handlers for it.
+*/
+package control
